backend/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
python-dotenv==1.0.0
Pillow
gunicorn==21.2.0
whitenoise==6.6.0
dj-database-url==2.1.0
psycopg[binary]
